rename: model 폴더 내 enum 클래스 enums 폴더로 이동
