+++ src/test/java/com/hhplus/tdd/concert/integrate/ConcertPaymentUseCaseIntegrationTest.java
@@ -3,7 +3,7 @@ package com.hhplus.tdd.concert.integrate;
 import com.hhplus.tdd.concert.application.usecase.ConcertPaymentUseCase;
 import com.hhplus.tdd.concert.domain.model.ConcertPayment;
 import com.hhplus.tdd.concert.domain.model.ConcertPaymentResult;
-import com.hhplus.tdd.concert.domain.model.PaymentStatus;
+import com.hhplus.tdd.concert.domain.model.enums.PaymentStatus;
 import com.hhplus.tdd.concert.domain.repository.ConcertPaymentRepository;
 import com.hhplus.tdd.concert.presentation.request.ConcertPaymentReq;
 import com.hhplus.tdd.waitingqueue.domain.model.WaitingQueue;
